+++ include/seqan3/io/sam_file/header.hpp
@@ -18,8 +18,8 @@
 #include <vector>
 
 #include <seqan3/alphabet/concept.hpp>
+#include <seqan3/alphabet/range/hash.hpp>
 #include <seqan3/io/sam_file/detail/cigar.hpp>
-#include <seqan3/range/hash.hpp>
 #include <seqan3/range/views/type_reduce.hpp>
 #include <seqan3/utility/type_traits/pre.hpp>
