+++ include/seqan3/range/hash.hpp
@@ -6,52 +6,14 @@
 // -----------------------------------------------------------------------------------------------------
 
 /*!\file
- * \author Enrico Seiler <enrico.seiler AT fu-berlin.de>
+ * \author [DEPRECATED] Enrico Seiler <enrico.seiler AT fu-berlin.de>
  * \brief Provides overloads for std::hash.
+ * \deprecated This header will be removed in 3.1.0; Please \#include <seqan3/alphabet/range/hash.hpp> instead.
  */
 
 #pragma once
 
-#include <seqan3/std/ranges>
+#include <seqan3/alphabet/range/hash.hpp>
 
-#include <seqan3/alphabet/hash.hpp>
-#include <seqan3/core/range/type_traits.hpp>
-
-namespace std
-{
-/*!\brief Struct for hashing a range of characters.
- * \ingroup alphabet
- * \tparam urng_t The type of the range; Must model std::ranges::input_range and the reference type of the range of the
-                  range must model seqan3::semialphabet.
- */
-template <ranges::input_range urng_t>
-//!\cond
-    requires seqan3::semialphabet<std::ranges::range_reference_t<urng_t>>
-//!\endcond
-struct hash<urng_t>
-{
-    /*!\brief Compute the hash for a range of characters.
-     * \tparam urng2_t  The same as `urng_t` (+- cvref); used to get forwarding reference in the interface.
-     * \param[in] range The input range to process. Must model std::ranges::input_range and the reference type of the
-                        range of the range must model seqan3::semialphabet.
-     * \returns size_t.
-     */
-    template <ranges::input_range urng2_t>
-    //!\cond
-        requires seqan3::semialphabet<std::ranges::range_reference_t<urng2_t>>
-    //!\endcond
-    size_t operator()(urng2_t && range) const noexcept
-    {
-        using alphabet_t = std::ranges::range_reference_t<urng_t>;
-        size_t result{0};
-        hash<alphabet_t> h{};
-        for (alphabet_t character : range)
-        {
-            result *= seqan3::alphabet_size<alphabet_t>;
-            result += h(character);
-        }
-        return result;
-    }
-};
-
-} // namespace std
+SEQAN3_DEPRECATED_HEADER(
+    "This header is deprecated and will be removed in SeqAn-3.1.0 Please #include <seqan3/alphabet/range/hash.hpp> instead.")
